requirements.txt
PyYAML>=3.12
termcolor>=1.1.0